Refactor, move editor and calendar components to common
