+++ src/app/components/editor-component.tsx
